requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
websockets==12.0
twilio==8.12.0

# Utilities
pydantic==2.6.1
python-dateutil==2.8.2
